+++ blocksuite/affine/shared/src/adapters/notion-text.ts
@@ -1,6 +1,4 @@
 import { DEFAULT_NOTE_BACKGROUND_COLOR } from '@blocksuite/affine-model';
-import { AdapterFactoryIdentifier } from '@blocksuite/affine-shared/adapters';
-import type { AffineTextAttributes } from '@blocksuite/affine-shared/types';
 import type { ExtensionType } from '@blocksuite/block-std';
 import { BlockSuiteError, ErrorCode } from '@blocksuite/global/exceptions';
 import type { DeltaInsert } from '@blocksuite/inline';
@@ -16,6 +14,9 @@ import {
   nanoid,
   type SliceSnapshot,
 } from '@blocksuite/store';
+
+import type { AffineTextAttributes } from '../types';
+import { AdapterFactoryIdentifier } from './types/adapter';
 
 type NotionEditingStyle = {
   0: string;
