+++ blocksuite/blocks/src/_common/adapters/extension.ts
@@ -1,18 +1,18 @@
 import {
+  AttachmentAdapterFactoryExtension,
   HtmlAdapterFactoryExtension,
+  ImageAdapterFactoryExtension,
   NotionHtmlAdapterFactoryExtension,
+  NotionTextAdapterFactoryExtension,
   PlainTextAdapterFactoryExtension,
 } from '@blocksuite/affine-shared/adapters';
 import type { ExtensionType } from '@blocksuite/block-std';
 
-import { AttachmentAdapterFactoryExtension } from './attachment.js';
 import { htmlInlineToDeltaMatchers } from './html/delta-converter/html-inline.js';
 import { inlineDeltaToHtmlAdapterMatchers } from './html/delta-converter/inline-delta.js';
-import { ImageAdapterFactoryExtension } from './image.js';
 import { MarkdownAdapterFactoryExtension } from './markdown/markdown.js';
 import { MixTextAdapterFactoryExtension } from './mix-text.js';
 import { notionHtmlInlineToDeltaMatchers } from './notion-html/delta-converter/html-inline.js';
-import { NotionTextAdapterFactoryExtension } from './notion-text.js';
 import { inlineDeltaToPlainTextAdapterMatchers } from './plain-text/delta-converter/inline-delta.js';
 
 export const AdapterFactoryExtensions: ExtensionType[] = [
