+++ blocksuite/blocks/src/_common/adapters/index.ts
@@ -1,7 +1,4 @@
-export * from './attachment.js';
 export * from './extension.js';
-export * from './image.js';
 export * from './markdown/index.js';
 export * from './mix-text.js';
 export * from './notion-html/index.js';
-export * from './notion-text.js';
